+++ RELEASE_NOTES.md
@@ -1,9 +1,12 @@
-# V.0.3.1
+# V.0.4.0
 
 
 ## Changes
-- No need GTF be sorted anymore.
-- Merge and Classify now support create index on the fly if not provided.
+- Added stats for isomatch classify
+- Output annotated GTF in isomatch classify
+- Added  `--guide-tss` and `--guide-tes` options for isomatch classify
+- Added toy example
+- improved documentation and help messages